Fix bug of networkHandler
*Bug* :
The middleware inscriptionNetworkHandler prevented to access gestion-personnes and gestion-machines, so impossible to sign in.

*solution* :
Make thing a little bit flexible by using django's resolver, to accept
either url name and url namespace.

*Warning* :

namespace should be used with caution cause they allowed every views in it.
It would be better to explicitely allow each urlname of views in apps...