Allow users to be unregistered in the subnet user
Web now check if the device is in the ldap to know if the user need to
be moved to the inscription zone. This is much more reliable (if the
ldap is stable).

In the future we might delete every occurence to the other subnets. The
only remaining should be the Internet-994 and the User-999.

See [Simplification des zones DHCP](https://git.resel.fr/resel/general/issues/1)
for more details