Add configuration for a new VM laputex-dev
This new VM is optionnal, and is activated only if the user do
`vagrant up laputex`. The choice made was to download the last repo
from /resel/laputex that way we don't need to keep the configuration
files in that repo.

In the future it would be good if a container exist which would make
testing much easier

Fixes: #52 Ajout de LaPuTeX dans l'env de test local