Delete campus redirection
Because the DNS master decided to do a round robin between doubidou and
skynet. We, developers of the ResEl website, had to do a workaround to
allow connections to both servers.

Thanks to this commit and the following :
 * ec98e4c7f272fb43c3ca66406e3b1aca0c8140c2 : Add REDIS caching and sessions
 * 4cbe18c0df423980ccac6a47cd871feba9d5adf4 : Create background task to update available ipset
 * e8c8059a0634f9b67453142192520ef825d1ff3d : Update CI to integrate new population functions
 * 256c36182a3bdd560a9346a998dc10e78e66d705 : Update get_free_ip function to take account for redis buffer
 * adb2636d6a4d327a25c0087b95d7d5b1f661dcd9 : Create command to save in redis the potential future user ips
 * b0aa618b5ad83c6194948af8be0028d72ef270cd : Create function to add ip on the redis server

Thus, the two servers can now work in HA (if the redis server is still
online and available).