Add campus bloc on home page
With this bloc the ResEl members will be able to have an overview of
what is happenning on the campus.

It is still a work in progress, and a revamp is not excluded. Also, all
of this **MUST** be cached before being put in production. For example
the club list, and the birthdays should be put in a daily cache. In this
light, this commit is at least blocked by:
https://git.resel.fr/resel/myresel/merge_requests/93